Renamed project from slashgif to ButBut
