+++ README.md
@@ -1,4 +1,4 @@
-slashgif
+ButBut
 ================
 
 Usage
@@ -12,5 +12,5 @@ URLs
 
 Misc
 ----
-* Build status: [![Build Status](https://travis-ci.org/TSAR-Industries/slashgif.svg?branch=master)](https://travis-ci.org/TSAR-Industries/slashgif)
+* Build status: [![Build Status](https://travis-ci.org/TSAR-Industries/ButBut.svg?branch=master)](https://travis-ci.org/TSAR-Industries/ButBut)
 * Banner created with [Taag](http://patorjk.com/software/taag)
