+++ application/pom.xml
@@ -3,18 +3,19 @@
 		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<parent>
-		<groupId>com.github.tsar-industries.slashgif</groupId>
-		<artifactId>slashgif-parent</artifactId>
+		<groupId>com.github.tsar-industries.butbut</groupId>
+		<artifactId>butbut-parent</artifactId>
 		<version>1.0-SNAPSHOT</version>
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>com.github.tsar-industries.slashgif.application</groupId>
+	<groupId>com.github.tsar-industries.butbut.application</groupId>
 	<artifactId>application</artifactId>
+	<name>ButBut - Application</name>
 
 	<dependencies>
 		<dependency>
-			<groupId>com.github.tsar-industries.slashgif.frontends.slack-integration</groupId>
+			<groupId>com.github.tsar-industries.butbut.frontends.slack-integration</groupId>
 			<artifactId>slack-integration</artifactId>
 			<scope>runtime</scope>
 		</dependency>
@@ -89,7 +90,7 @@
 										implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
 								<transformer
 										implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-									<mainClass>com.github.tsar_industries.slashgif.application.SlashgifApplication</mainClass>
+									<mainClass>com.github.tsar_industries.butbut.application.ButButApplication</mainClass>
 								</transformer>
 							</transformers>
 						</configuration>
@@ -98,5 +99,4 @@
 			</plugin>
 		</plugins>
 	</build>
-
 </project>
