+++ application/src/main/java/com/github/tsar_industries/butbut/application/ButbutBundle.java
@@ -1,4 +1,4 @@
-package com.github.tsar_industries.slashgif.application;
+package com.github.tsar_industries.butbut.application;
 
 import io.dropwizard.ConfiguredBundle;
 import io.dropwizard.java8.Java8Bundle;
@@ -9,9 +9,9 @@ import io.dropwizard.setup.Environment;
  * Bundles are basically plugins in dropwizard <br/>
  * This one gets automatically found by guice and registered in dropwizard
  *
- * @see SlashgifApplication#getGuiceBundle(Bootstrap)
+ * @see ButButApplication#getGuiceBundle(Bootstrap)
  */
-public class SlashgifBundle implements ConfiguredBundle<SlashgifConfiguration>
+public class ButButBundle implements ConfiguredBundle<ButButConfiguration>
 {
 	@Override
 	public void initialize(Bootstrap<?> bootstrap)
@@ -20,7 +20,7 @@ public class SlashgifBundle implements ConfiguredBundle<SlashgifConfiguration>
 	}
 
 	@Override
-	public void run(SlashgifConfiguration configuration, Environment environment) throws Exception
+	public void run(ButButConfiguration configuration, Environment environment) throws Exception
 	{
 
 	}
