+++ application/src/main/java/com/github/tsar_industries/butbut/application/ButbutConfiguration.java
@@ -1,0 +1,10 @@
+package com.github.tsar_industries.butbut.application;
+
+import io.dropwizard.Configuration;
+
+/**
+ * This is the java-representation of the yaml-config file
+ */
+public class ButButConfiguration extends Configuration
+{
+}
