+++ frontends/api/pom.xml
@@ -4,11 +4,12 @@
 		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<parent>
 		<artifactId>frontends</artifactId>
-		<groupId>com.github.tsar-industries.slashgif.frontends</groupId>
+		<groupId>com.github.tsar-industries.butbut.frontends</groupId>
 		<version>1.0-SNAPSHOT</version>
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>com.github.tsar-industries.slashgif.frontends.api</groupId>
+	<groupId>com.github.tsar-industries.butbut.frontends.api</groupId>
 	<artifactId>api</artifactId>
+	<name>ButBut - Frontends Api</name>
 </project>
