+++ frontends/pom.xml
@@ -3,14 +3,15 @@
 		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<parent>
-		<groupId>com.github.tsar-industries.slashgif</groupId>
-		<artifactId>slashgif-parent</artifactId>
+		<groupId>com.github.tsar-industries.butbut</groupId>
+		<artifactId>butbut-parent</artifactId>
 		<version>1.0-SNAPSHOT</version>
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>com.github.tsar-industries.slashgif.frontends</groupId>
+	<groupId>com.github.tsar-industries.butbut.frontends</groupId>
 	<artifactId>frontends</artifactId>
+	<name>ButBut - Frontends</name>
 	<packaging>pom</packaging>
 	<modules>
 		<module>api</module>
