+++ search-providers/api/pom.xml
@@ -3,17 +3,18 @@
 		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<parent>
-		<groupId>com.github.tsar-industries.slashgif.search-providers</groupId>
+		<groupId>com.github.tsar-industries.butbut.search-providers</groupId>
 		<artifactId>search-providers</artifactId>
 		<version>1.0-SNAPSHOT</version>
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>com.github.tsar-industries.slashgif.search-providers.api</groupId>
+	<groupId>com.github.tsar-industries.butbut.search-providers.api</groupId>
 	<artifactId>api</artifactId>
+	<name>ButBut - Search Providers Api</name>
 	<dependencies>
 		<dependency>
-			<groupId>com.github.tsar-industries.slashgif.service-layer.api</groupId>
+			<groupId>com.github.tsar-industries.butbut.service-layer.api</groupId>
 			<artifactId>api</artifactId>
 		</dependency>
 	</dependencies>
