+++ search-providers/google-image-search/src/main/java/com/github/tsar_industries/butbut/search_providers/google_image_search/GoogleSearchProvider.java
@@ -1,7 +1,7 @@
-package com.github.tsar_industries.slashgif.search_providers.google_image_search;
+package com.github.tsar_industries.butbut.search_providers.google_image_search;
 
-import com.github.tsar_industries.slashgif.service_layer.api.Image;
-import com.github.tsar_industries.slashgif.search_providers.api.SearchProviderApi;
+import com.github.tsar_industries.butbut.service_layer.api.Image;
+import com.github.tsar_industries.butbut.search_providers.api.SearchProviderApi;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
