+++ search-providers/pom.xml
@@ -3,14 +3,15 @@
 		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<parent>
-		<groupId>com.github.tsar-industries.slashgif</groupId>
-		<artifactId>slashgif-parent</artifactId>
+		<groupId>com.github.tsar-industries.butbut</groupId>
+		<artifactId>butbut-parent</artifactId>
 		<version>1.0-SNAPSHOT</version>
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>com.github.tsar-industries.slashgif.search-providers</groupId>
+	<groupId>com.github.tsar-industries.butbut.search-providers</groupId>
 	<artifactId>search-providers</artifactId>
+	<name>ButBut - Search Providers</name>
 	<packaging>pom</packaging>
 
 	<modules>
